backend/pyproject.toml
[tool.poetry]
name = "life-navigator-backend"
version = "0.1.0"
description = "Life Navigator FastAPI Backend - Multi-tenant HIPAA-compliant platform"
authors = ["Life Navigator Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
# Security-critical updates
fastapi = "^0.115.0"  # Updated to fix DoS vulnerabilities
python-multipart = "^0.0.18"  # Updated to fix DoS vulnerability
sqlalchemy = "^2.0.36"  # Updated to fix SQL injection vulnerability
pillow = "^11.0.0"  # Updated for security
cryptography = "^46.0.3"  # Updated for security

# Compatibility updates for fastapi 0.115
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"

# Keep existing versions to avoid dependency conflicts
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}  # Updated to fix CRITICAL vulnerabilities
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^2.1.0"
redis = "^5.0.1"
celery = "^5.3.6"
httpx = "^0.26.0"
tenacity = "^8.2.3"
python-dotenv = "^1.0.0"
structlog = "^24.1.0"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.40.0"}
slowapi = "^0.1.9"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
pgvector = "^0.2.4"
numpy = "^1.26.3"
plaid-python = "^18.0.0"
stripe = "^7.11.0"
twilio = "^8.13.0"
# Resend for transactional emails (modern, simpler than SendGrid)
# httpx already included above for async HTTP
boto3 = "^1.34.27"
google-cloud-storage = "^2.14.0"
google-cloud-secret-manager = "^2.18.1"
python-magic = "^0.4.27"

# Knowledge Graph and Vector Storage
neo4j = "^5.14.0"  # Neo4j async driver for graph database
qdrant-client = "^1.7.0"  # Qdrant vector database client
# openai removed - use internal GraphRAG for embeddings

# Supabase - Primary Database Client
supabase = "^2.0.0"  # Supabase Python client for auth and database operations

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"
black = "^25.9.0"
ruff = "^0.1.13"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
faker = "^22.5.1"
openai = "^1.6.0"  # Dev only - for local experiments, NOT production

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "sqlalchemy.*",
    "alembic.*",
    "redis.*",
    "celery.*",
    "neo4j.*",
    "qdrant_client.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=app --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
